// src/types/generated/contract.ts
export interface LayerRule {
  anchor?: string | null;
  exclude?: MatchRules | null;
  "keyboard-interactivity"?: string | null;
  layer?: string | null;
  margin?: Margin | null;
  "match-rules": MatchRules;
  namespace: string;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "MatchRules".
 */
export interface MatchRules {
  "app-id"?: string | null;
  "at-startup"?: boolean | null;
  "is-active"?: boolean | null;
  "is-floating"?: boolean | null;
  "is-focused"?: boolean | null;
  "is-fullscreen"?: boolean | null;
  "is-maximized"?: boolean | null;
  namespace?: string | null;
  title?: string | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "Margin".
 */
export interface Margin {
  bottom?: number | null;
  left?: number | null;
  right?: number | null;
  top?: number | null;
}